+++ tests/testapp/models.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import, unicode_literals
 
 from django.db import models
-from django.db.models import Q, Lookup
+from django.db.models import Lookup
 from django.db.models.fields import Field
 from django.utils.encoding import python_2_unicode_compatible
 from mptt.fields import TreeForeignKey
@@ -9,6 +9,7 @@ from mptt.models import MPTTModel
 from treebeard.mp_tree import MP_Node
 from treebeard.ns_tree import NS_Node
 
+from relativity.compat import Q
 from relativity.fields import L, Relationship
 from relativity.mptt import MPTTDescendants, MPTTSubtree
 from relativity.treebeard import MP_Descendants, NS_Descendants, MP_Subtree, NS_Subtree