+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Allowed for callable predicates (thanks django-reverse-unique)
 - Fix bug in get_forward_related_filter() causing unnecessarily complicated queries
 - Fix bug when building exclude filters
+- Fix migrations under Django <2.0
 
 ## 0.1.3 - 2018-08-31
 - Added descendant and subtree fields for django-treebeard